fix(test-helpers-instruction): prefix list should not contain optionals
When prefix lists are created via a generator that there must be a
way for the generator to create an empty prefix list. Currently this
is done by having the NO_PREFIX prefix be an optional without a
value, since any prefix object would be converted to a byte array
adding bytes to the instruction, which the NO_PREFIX prefix should
not do. This means that the prefix list passed when constructing an
Instruction  may contain the NO_PREFIX prefix, and it should do
nothing. If it is the only prefix in the list then the list is
treated as though it has no  members.
The prefixes being option in the prefix list is a result of how they
are generated, but for the purposes of the Instruction this
information is redundant since an empty vector has the same result.
By copying the prefixes that are relevant into the instruction prefix
list we can fix a conceptual error and a compile error.
The conceptual error being that there is such a think as a non
existent prefix from the point of view of an instruction.
The compile error being that the Parameter.toString method takes
std::vector<prefix_t> and not std::vector<std::optional<prefix_t>>.