add(test-helpers): view that gives cartesian product of input containers
It is sometimes needed to iterate over a set of containers in such a
way that that you visit every element in the cartesian product of the
containers of the set. This view helps with that by iterating though
the containers and return a tuple of values such that each tuple has
exactly one value from each container, and each element in the
cartesian product of the containers is returned once.