refactor(test): use the new PrefixListGenerator
Update Instruction and InstructionGenerator to use a vector of
prefixes instead of a vector of optional prefixes.

Fix generatePrefixList and PrefixListGenerator to move their
arguments instead of copy them. This is required to take other
generators which can't be copied.

Add PREFIXES_DETAIL_::makePrefixListGenerator with overloads to
allow the PREFIXES_ macro to construct the correct prefix list
generator. The PrefixListGenerator will not create empty prefix lists
and so if an empty prefix is required we need to pass it separately.
The NO_PREFIX macro can be passed as the first argument to the
PREFIXES_ macro to specify that an empty prefix list should also be
generated. If it appears anywhere else in the list the
PrefixListGenerator will ignore it (this is needed to allow the
IF_X64_DISASSEMBLER_ to work).

Change the definition of prefix macros to not be generators but
instead just raw lists.

Update ON_X64_ macro see the prefix list as a vector of prefixes and
not a vector of optional prefixes.

Update IF_X64_DISASSEMBLER_ macro lambda to return an optional
generator so that its return value can be passed to
generatePrefixList in the PREFIXES_ macro.