refactor(test-helpers): change internals of conditional operands
Currently, the conditional operands rely on complex lambda functions
to determine the mode and if that mode is relevant to the operand.
This has issues in that it tries to combine both address and
parameter modes into a single result, it is difficult to change, and
it results in the creation of many lambda objects. Instead, the body
of these lambdas can be split into two functions, one for address
relevant operands and one for parameter relevant operands. These
functions are easier to maintain and don't require complicated logic
to combine both types of operand size.

There was also an error when selecting an X64_PARAMETER_MODE in that
it was not selected if there was a OPERAND_SIZE_OVERRIDE prefix even
if the REX prefix was present.

Add IF_ macro to allow simplifying conditional operand statements which
only have two options.